+++ modules/actor/sheet/cpr-container-sheet.js
@@ -1,10 +1,12 @@
-/* global mergeObject $ game getProperty duplicate */
+/* global mergeObject $ game getProperty duplicate setProperty */
 import CPRActorSheet from "./cpr-actor-sheet.js";
 import LOGGER from "../../utils/cpr-logger.js";
 import SystemUtils from "../../utils/cpr-systemUtils.js";
 import CPRChat from "../../chat/cpr-chat.js";
 import CPRItem from "../../item/cpr-item.js";
 import PurchasePartPrompt from "../../dialog/cpr-purchase-part-prompt.js";
+import ConfigureSellToPrompt from "../../dialog/cpr-container-configure-sell-to-prompt.js";
+import PurchaseOrderPrompt from "../../dialog/cpr-container-vendor-purchase-order-prompt.js";
 
 /**
  * Implement the sheet for containers and shop keepers. This extends CPRActorSheet to make use
@@ -73,10 +75,11 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
     html.find(".container-type-dropdown").change((event) => this._setContainerType(event));
     // Toggle the state of a flag for the data of the checkbox
     html.find(".checkbox-toggle").click((event) => this._checkboxToggle(event));
-
     // Eurobucks management
     html.find(".eurobucks-input-button").click((event) => this._updateEurobucks(event));
     html.find(".eurobucks-open-ledger").click(() => this.actor.showLedger());
+    // Configure container to purchase items from players
+    html.find(".vendor-configure-sell-to").click(() => this._configureSellTo());
 
     super.activateListeners(html);
   }
@@ -219,7 +222,7 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
       SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.insufficientPermissions"));
       return;
     }
-    console.log(this);
+
     const transferredItemData = duplicate(item.data);
     let cost = 0;
     if (item.type === "ammo" && item.data.data.variety !== "grenade" && item.data.data.variety !== "rocket") {
@@ -254,15 +257,21 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
       const username = game.user.name;
       let reason = "";
       if (amount > 1) {
-        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.multiple",
+        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.multiplePurchased",
           { amount, name: item.name, price: cost })} - ${username}`;
       } else {
-        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.single",
+        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.singlePurchased",
           { name: item.name, price: cost })} - ${username}`;
       }
+      const vendorReason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.vendorSold",
+        {
+          name: item.name,
+          quantity: amount,
+          purchaser: tradePartnerActor.name,
+          price: cost,
+        })} - ${username}`;
       await tradePartnerActor.deltaLedgerProperty("wealth", -1 * cost, reason);
-      itemName, price, quantity, reason, purchaser, seller
-      await this.actor.recordTransaction(item.name, cost, 1, tradePartnerActor, this.actor);
+      await this.actor.recordTransaction(cost, vendorReason);
     }
     if (tradePartnerActor.automaticallyStackItems(new CPRItem(transferredItemData))) {
       await tradePartnerActor.createEmbeddedDocuments("Item", [transferredItemData]);
@@ -282,67 +291,72 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
    * the tradePartner and money is deducted accordingly.
    *
    * @private
+   * @param {string} sellerId - actor.id of the person selling the item
    * @param {Item} item - object to be purchased
-   * @param {boolean} all - Toggle to purchase all of the items in the stack or just a part of them
    */
-  async _sellItem(item, all) {
-    LOGGER.trace("_sellItem | CPRContainerSheet | Called.");
-    if (this.tradePartnerId === undefined || this.tradePartnerId === "") {
-      SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradeWithWarn"));
+  async _sellItemTo(event) {
+    LOGGER.trace("_sellItemTo | CPRContainerSheet | Called.");
+
+    // Players must have Owned permission on Containers for them to function properly
+    if (!this.actor.isOwner) {
+      SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.insufficientPermissions"));
       return;
     }
-    const transferredItemData = duplicate(item.data);
+    const dragData = JSON.parse(event.dataTransfer.getData("text/plain"));
+    const tradePartnerActor = game.actors.get(dragData.actorId);
+
+    const itemData = duplicate(dragData.data);
+    const amount = parseInt(itemData.data.amount, 10);
+    const vendorData = this.actor.data;
+    const vendorConfig = vendorData.data.vendor;
+    const username = game.user.name;
+
     let cost = 0;
-    if (item.type === "ammo" && item.data.data.variety !== "grenade" && item.data.data.variety !== "rocket") {
+    if (itemData.type === "ammo" && itemData.data.variety !== "grenade" && itemData.data.variety !== "rocket") {
       // Ammunition, which is neither grenades nor rockets, are prices are for 10 of them (pg. 344)
-      cost = item.data.data.price.market / 10;
+      cost = parseInt(parseInt(itemData.data.price.market, 10) / 10, 10);
     } else {
-      cost = item.data.data.price.market;
+      cost = parseInt(itemData.data.price.market, 10);
     }
-    if (!all) {
-      const itemText = SystemUtils.Format("CPR.dialog.purchasePart.text", { itemName: item.name });
-      const formData = await PurchasePartPrompt.RenderPrompt(itemText).catch((err) => LOGGER.debug(err));
-      if (formData === undefined) {
-        return;
-      }
-      const newAmount = parseInt(formData.purchaseAmount, 10);
-      if (newAmount < 1 || newAmount >= parseInt(item.data.data.amount, 10)) {
-        SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.dialog.purchasePart.wrongAmountWarning"));
-        return;
-      }
-      transferredItemData.data.amount = newAmount;
-      cost *= newAmount;
-    } else {
-      cost *= item.data.data.amount;
-    }
-    const tradePartnerActor = game.actors.get(this.tradePartnerId);
-    if (!getProperty(this.actor.data, "flags.cyberpunk-red-core.items-free")) {
-      if (tradePartnerActor.data.data.wealth.value < cost) {
-        SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradePriceWarn"));
-        return;
-      }
-      const { amount } = transferredItemData.data;
-      const username = game.user.name;
+    const percent = parseInt(vendorConfig.itemTypes[itemData.type].purchasePercentage, 10);
+
+    let vendorOffer = parseInt(((amount * cost * percent) / 100), 10);
+    vendorOffer = Math.min(vendorOffer, vendorData.data.wealth.value);
+
+    const offerMessage = `${SystemUtils.Format("CPR.dialog.container.vendor.offerToBuy",
+      {
+        vendorName: tradePartnerActor.name,
+        vendorOffer,
+        itemName: itemData.name,
+        percent,
+      })}`;
+    const formData = await PurchaseOrderPrompt.RenderPrompt(offerMessage).catch((err) => LOGGER.debug(err));
+
+    if (formData !== undefined) {
+      await this.actor.createEmbeddedDocuments("Item", [itemData]).then(tradePartnerActor.deleteEmbeddedDocuments("Item", [itemData._id]));
+
       let reason = "";
       if (amount > 1) {
-        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.multiple",
-          { amount, name: item.name, price: cost })} - ${username}`;
+        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.multipleSold",
+          {
+            amount,
+            name: itemData.name,
+            price: vendorOffer,
+            vendor: this.actor.name,
+          })} - ${username}`;
       } else {
-        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.single",
-          { name: item.name, price: cost })} - ${username}`;
+        reason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.singleSold",
+          { name: itemData.name, price: vendorOffer, vendor: this.actor.name })} - ${username}`;
       }
-      await tradePartnerActor.deltaLedgerProperty("wealth", -1 * cost, reason);
-    }
-    if (tradePartnerActor.automaticallyStackItems(new CPRItem(transferredItemData))) {
-      await tradePartnerActor.createEmbeddedDocuments("Item", [transferredItemData]);
-    }
-    if (!getProperty(this.actor.data, "flags.cyberpunk-red-core.infinite-stock")) {
-      if (all) {
-        await this._deleteOwnedItem(item, true);
-      } else {
-        const keepAmount = item.data.data.amount - transferredItemData.data.amount;
-        await this.actor.updateEmbeddedDocuments("Item", [{ _id: item.id, "data.amount": keepAmount }]);
-      }
+      const vendorReason = `${SystemUtils.Format("CPR.containerSheet.tradeLog.vendorPurchased",
+        {
+          name: itemData.name,
+          quantity: itemData.data.amount,
+          seller: tradePartnerActor.name,
+          price: vendorOffer,
+        })} - ${username}`;
+      await tradePartnerActor.deltaLedgerProperty("wealth", vendorOffer, reason);
+      await this.actor.recordTransaction(vendorOffer, vendorReason);
     }
   }
 
@@ -413,10 +427,64 @@ export default class CPRContainerActorSheet extends CPRActorSheet {
   async _onDrop(event) {
     LOGGER.trace("_onDrop | CPRContainerSheet | Called.");
     const playersCanCreate = getProperty(this.actor.data, "flags.cyberpunk-red-core.players-create");
-    if (game.user.isGM || playersCanCreate) {
+    const playersCanSell = getProperty(this.actor.data, "flags.cyberpunk-red-core.players-sell");
+    if (game.user.isGM || playersCanCreate || playersCanSell) {
+      const dragData = JSON.parse(event.dataTransfer.getData("text/plain"));
+      if (playersCanSell) {
+        const vendorData = this.actor.data.data.vendor;
+        if (dragData.type === "Item") {
+          const itemData = dragData.data;
+          if (typeof vendorData.itemTypes[itemData.type] !== "undefined" && vendorData.itemTypes[itemData.type].isPurchasing) {
+            await this._sellItemTo(event);
+            return;
+          }
+          SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradeDragNotBuying"));
+          return;
+        }
+      }
       super._onDrop(event);
     } else {
       SystemUtils.DisplayMessage("warn", SystemUtils.Localize("CPR.messages.tradeDragInWarn"));
+    }
+  }
+
+  /**
+   * Configure Container Actor to be able to purchase items from a player.
+   *
+   * @private
+   * @callback
+   */
+  async _configureSellTo() {
+    LOGGER.trace("_configureSellTo | CPRContainerSheet | Called.");
+    const actorData = duplicate(this.actor.data);
+    const promptData = {};
+    promptData.itemTypes = [];
+    const itemEntities = game.system.template.Item;
+    game.system.template.Item.types.forEach((itemType) => {
+      if (itemEntities[itemType].templates.includes("physicalItem")) {
+        promptData.itemTypes.push(itemType);
+      }
+    });
+    promptData.currentConfig = actorData.data.vendor;
+
+    promptData.itemTypes.forEach((itemType) => {
+      if (typeof promptData.currentConfig.itemTypes[itemType] === "undefined") {
+        promptData.currentConfig.itemTypes[itemType] = {
+          isPurchasing: false,
+          purchasePercentage: 0,
+        };
+      }
+    });
+
+    const formData = await ConfigureSellToPrompt.RenderPrompt(promptData).catch((err) => LOGGER.debug(err));
+    if (formData !== undefined) {
+      promptData.itemTypes.forEach((itemType) => {
+        const isPurchasing = formData[`sell-${itemType}`];
+        const purchasePercentage = (isPurchasing) ? formData[`pct-${itemType}`] : 0;
+        promptData.currentConfig.itemTypes[itemType] = { isPurchasing, purchasePercentage };
+      });
+      setProperty(actorData, "data.vendor", promptData.currentConfig);
+      this.actor.update(actorData);
     }
   }
 }